+++ packages/synthetics-sdk-broken-links/src/link_utils.ts
@@ -12,21 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { HTTPResponse } from 'puppeteer';
+import { Browser, HTTPResponse, Page } from 'puppeteer';
 import {
-  ResponseStatusCode,
-  ResponseStatusCode_StatusClass,
   BrokenLinksResultV1,
   BrokenLinksResultV1_BrokenLinkCheckerOptions,
   BrokenLinksResultV1_BrokenLinkCheckerOptions_LinkOrder,
-  BrokenLinksResultV1_SyntheticLinkResult,
   BrokenLinksResultV1_BrokenLinkCheckerOptions_PerLinkOption,
+  BrokenLinksResultV1_SyntheticLinkResult,
+  ResponseStatusCode,
+  ResponseStatusCode_StatusClass,
   SyntheticResult,
 } from '@google-cloud/synthetics-sdk-api';
 import {
   BrokenLinkCheckerOptions,
-  StatusClass,
   LinkOrder,
+  StatusClass,
 } from './broken_links';
 
 /**
@@ -126,11 +126,17 @@ export function checkStatusPassing(
  * @returns The sanitized input options if validation passes.
  * @throws {Error} If any of the input options fail validation.
  */
-export function validateInputOptions(inputOptions: BrokenLinkCheckerOptions) {
+export function validateInputOptions(
+  inputOptions: BrokenLinkCheckerOptions
+): BrokenLinkCheckerOptions {
   if (!inputOptions.origin_url) {
     throw new Error('Missing origin_url in options');
-  } else if (!inputOptions.origin_url.startsWith('http')) {
-    throw new Error('origin_url must start with `http`');
+  } else if (
+    typeof inputOptions.origin_url !== 'string' ||
+    (!inputOptions.origin_url.startsWith('http') &&
+      !inputOptions.origin_url.endsWith('.html'))
+  ) {
+    throw new Error('origin_url must be a string that starts with `http`');
   }
 
   // check link_limit
@@ -300,10 +306,13 @@ export function setDefaultOptions(
   >;
   for (const optionName of optionsKeys) {
     // per_link_options and linkorder are handled below
+    if (optionName === 'per_link_options' || optionName === 'link_order')
+      continue;
+
     if (
       !(optionName in inputOptions) ||
-      optionName === 'per_link_options' ||
-      optionName === 'link_order'
+      // eslint-disable-next-line  @typescript-eslint/no-explicit-any
+      (inputOptions as any)[optionName] === undefined
     ) {
       // eslint-disable-next-line  @typescript-eslint/no-explicit-any
       (outputOptions as any)[optionName] = defaulOptions[optionName];
@@ -521,4 +530,82 @@ export function createSyntheticResult(
   };
 
   return synthetic_result;
+}
+
+/**
+ * Opens a new Puppeteer page within the provided browser instance, disables caching, and returns the created page.
+ *
+ * @param browser - The Puppeteer browser instance in which to open a new page.
+ * @returns A Promise that resolves with the newly created Puppeteer page or
+ *          rejects if an error occurs during page creation.
+ * @throws {Error} If an error occurs while opening a new page, it throws an
+ *                 error with an appropriate message.
+ */
+export async function openNewPage(browser: Browser) {
+  try {
+    const page = await browser.newPage();
+    page.setCacheEnabled(false);
+    return page;
+  } catch (pageError) {
+    if (pageError instanceof Error) process.stderr.write(pageError.message);
+    throw new Error('An error occurred while opening a new puppeteer.Page.');
+  }
+}
+
+/**
+ * Closes the provided Puppeteer browser instance and handles any errors
+ * gracefully. No error is thrown as even if this errors we do not need to fail
+ * the entire execution as Cloud Functions will handle the cleanup.
+ *
+ * @param browser - The Puppeteer browser instance to close.
+ */
+export async function closeBrowser(browser: Browser) {
+  try {
+    await browser.close();
+  } catch (err) {
+    if (err instanceof Error) process.stderr.write(err.message);
+  }
+}
+
+/**
+ * Closes the provided Puppeteer pages handles any errors
+ * gracefully. No error is thrown as even if this errors we do not need to fail
+ * the entire execution as Cloud Functions will handle the cleanup.
+ *
+ * @param browser - The Puppeteer browser instance to close.
+ */
+export async function closePagePool(pagePool: Page[]) {
+  try {
+    // Close all pages in the pool
+    await Promise.all(pagePool.map(async (page) => await page.close()));
+  } catch (err) {
+    if (err instanceof Error) process.stderr.write(err.message);
+  }
+}
+
+/**
+ * If the `link_order` is set to `RANDOM`, the links will be shuffled randomly.
+ * Otherwise, the links will be copied without shuffling. Truncate to
+ * `link_limit` regardless
+ *
+ * @param links - The array of links to process.
+ * @param link_limit - The maximum number of links to retain.
+ * @param link_order - Whether or not to shuffle links (enum value).
+ * @returns A new array of links that have been truncated based on the `link_limit`.
+ */
+export function shuffleAndTruncate(
+  links: LinkIntermediate[],
+  link_limit: number,
+  link_order: BrokenLinksResultV1_BrokenLinkCheckerOptions_LinkOrder
+): LinkIntermediate[] {
+  // shuffle links if link_order is `RANDOM` and truncate to link_limit
+
+  // Shuffle the links if link_order is RANDOM, or copy the original array
+  const linksToFollow =
+    link_order === BrokenLinksResultV1_BrokenLinkCheckerOptions_LinkOrder.RANDOM
+      ? [...links].sort(() => Math.random() - 0.5)
+      : [...links];
+
+  // Truncate the processed array to match the link_limit
+  return linksToFollow.slice(0, link_limit! - 1);
 }
